feat: expose useMarkdown composable
